Improved component removal
I do actual placement new now, using a buffer abstraction for components.
This way component's default ctr won't be called unless required. So won't be the destructors.
When removing data, the last component is MOVED (or copied if no move constructor is present) into the removed one. The removed component is destructed and so is the moved one (after the move or the copy).