Fixed a major bug in EntityIterator
EntityIterator was fetching the entity count at every call of ++, begin() or end().
If we added a new entity while iterating, it would screw things over.
I am sure it would potentially create problems if we go over the limit of entites (hence go to the next generation of entities for an index) or if we fetch from free list, but it's better than before